pyproject.toml
[project]
name = "rag-knowledge-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.76",
    "celery>=5.6.2",
    "cohere>=5.20.7",
    "fastapi>=0.135.2",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "llama-index>=0.14.19",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-anthropic>=0.11.1",
    "llama-index-vector-stores-pinecone>=0.8.0",
    "pinecone[grpc]>=7.3.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "redis>=7.4.0",
    "uvicorn>=0.42.0",
    "voyageai>=0.3.7",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]
